requirements.txt
jupyter_client==8.6.3
jupyter_console==6.6.3
jupyterlab==4.3.5
laplace_torch==0.2.1
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
portion==2.6.0
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
scikit-image==0.25.2
tabulate==0.9.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
